fix desktop file modifications (#435)
We want to add a StartupWMClass entry for the Chrome desktop file, to
prevent multiple icons from being displayed at once when the different
"modes" of Chrome are in use (standard, new window, incognito). At some
point in the recent past, the section titles were changed. This
sometimes caused a bug when adding the "StartupWMClass" entry to the
desktop file; it is undetermined if the source of the bug was the
upgrade/reinstallation, or the "install.sh" script itself. This patch
introduces a better practice for conditionally adding the
"StartupWMClass" entries to the different sections, and reports if a
section is not found (which is useful for debugging).

Relates to gh-431