apply changes requested in review
- remove unecessary abstraction of simple dnf commands. when fedy
interfaces directly with dnf to install/remove a package and nothing
else is necessary (this is the ideal workflow), we should highly prefer
to simply use metadata.json as the go-to for defining those actions.

- remove overriding the value of dnf's clean_requires_on_remove
configuration setting. this defaults to True (and likely always has
since introduction), so by overriding it, all we are doing is ignoring
user's configuration of dnf -- this is not ideal.