Switch to using dnf groups for codecs
Fedora groups are maintained here:
https://pagure.io/fedora-comps

RPM Fusion groups are maintained here
https://github.com/rpmfusion-infra/rpmfusion-free-comps
https://github.com/rpmfusion-infra/rpmfusion-nonfree-comps

Thoses are better way to complement the needed codecs because they
should install if the package is already installed in the user system

For example it will install gstreamer-ffmpeg if gstreamer(-0.10) is
installed and gstreamer1-libav if gstreamer1 is installed.

Others cases might be handled in the future

As it's not easy to check for group availability, the check assume
gstreamer1-libav. If the package was already installed, then it"s only
a matter to uninstall it and re-install the whole group.

Signed-off-by: Nicolas Chauvet <kwizart@gmail.com>